remove logger due to multiprocessing; fix file cache sync

+++ tools/disney2luisa.py
@@ -8,8 +8,7 @@ import re
 import multiprocessing
 import time
 
-# from colorama import Fore, Style
-import logging
+from colorama import Fore, Style
 
 
 class ConverterInfo:
@@ -57,25 +56,8 @@ class ConverterInfo:
             raise Exception('Multiple environment names found')
 
 
-# def log_info(msg, *args, **kwargs):
-#     print(f'{Fore.GREEN}[INFO]{Style.RESET_ALL} {msg}', *args, **kwargs)
-
-
-global logger
-def logger_init():
-    log_format = '[%(asctime)s] [%(levelname)s] %(message)s'
-    date_format = '%Y-%m-%d %H:%M:%S'
-    formatter = logging.Formatter(fmt=log_format, datefmt=date_format)
-
-    logging.basicConfig(level=logging.DEBUG, format=log_format, datefmt=date_format)
-
-    global logger
-    logger = logging.getLogger(__file__)
-
-    # file_handler = logging.FileHandler(log_file)
-    # file_handler.setFormatter(formatter)
-    # file_handler.setLevel(logging.DEBUG)
-    # logger.addHandler(file_handler)
+def log_info(msg, *args, **kwargs):
+    print(f'{Fore.GREEN}[INFO]{Style.RESET_ALL} {msg}', *args, **kwargs)
 
 
 def split_obj(path: Path, output_dir: Path) -> dict:
@@ -163,7 +145,7 @@ def split_obj(path: Path, output_dir: Path) -> dict:
     def move_to_end_of(src: str, dst: str):
         input_path = output_dir / f'{src}.obj'
         output_path = output_dir / f'{dst}.obj'
-        # logger.info(f'Trying to move {input_path} to {output_path}')
+        # log_info(f'Trying to move {input_path} to {output_path}')
         if input_path == output_path:
             return
         if not output_path.exists():
@@ -186,10 +168,11 @@ def split_obj(path: Path, output_dir: Path) -> dict:
         geo_index_nums = {}
         geo_lines = {}
         geo_lines.setdefault('default', [])
-        geo2material = {
+        file2material = {
             # 'geo_isBeach_sand': 'mat_isBeach_sand',
         }
         geo_name = None
+        file_name = None
         material = None
         geo_name2file_name = {
             'default': 'default',
@@ -199,47 +182,54 @@ def split_obj(path: Path, output_dir: Path) -> dict:
         for line in f:
             if line.startswith('g '):
                 geo_name_new = line[2:].strip()
+                file_name_new = geo_name_new
                 material = None
-                write_geo_line(line, geo_name_new, geo_lines)
                 if g_mode:
                     g_mode = False
                     index_num = index_next - index
                     index = index_next.copy()
+                    write_geo_force(file_name, geo_lines)
                 elif geo_name is not None:
                     g_mode = True
-                    write_geo_force('default', geo_lines)
-                    move_to_end_of('default', geo_name_new)
+                    write_geo_force(file_name, geo_lines)
+                    move_to_end_of(file_name, file_name_new)
                 geo_name = geo_name_new
+                file_name = geo_name_new
+                write_geo_line(line, file_name, geo_lines)
             elif line.startswith('f '):
                 if material is None:
-                    geo_name = geo_name2file_name[geo_name]
-                    material = geo2material[geo_name]
+                    file_name_new = geo_name2file_name[geo_name]
+                    material = file2material[file_name_new]
+                    write_geo_force(file_name, geo_lines)
+                    move_to_end_of(file_name, file_name_new)
+                    merge_geo(geo_name, index_num, geo_index_nums)
+                    file_name = file_name_new
                 face = reindex(line, index - geo_index_nums[geo_name])
-                write_geo_line(f'f {face}\n', geo_name, geo_lines)
+                write_geo_line(f'f {face}\n', file_name, geo_lines)
             elif line.startswith('usemtl '):
                 material = line[7:].strip()
-                geo_name2file_name[geo_name] = f'geo_{path.stem}_{material}'
+                file_name = f'geo_{path.stem}_{material}'
+                geo_name2file_name[geo_name] = file_name
                 material = f'mat_{path.stem}_{material}'
-                move_to_end_of(geo_name, geo_name2file_name[geo_name])
-                geo_name = geo_name2file_name[geo_name]
-                geo2material[geo_name] = material
+                write_geo_force(geo_name, geo_lines)
+                move_to_end_of(geo_name, file_name)
                 merge_geo(geo_name, index_num, geo_index_nums)
-                write_geo_line(line, geo_name, geo_lines)
+                file2material[file_name] = material
+                write_geo_line(line, file_name, geo_lines)
             elif line.startswith('v '):
                 index_next.v_index += 1
-                write_geo_line(line, geo_name, geo_lines)
+                write_geo_line(line, file_name, geo_lines)
             elif line.startswith('vn '):
                 index_next.vn_index += 1
-                write_geo_line(line, geo_name, geo_lines)
+                write_geo_line(line, file_name, geo_lines)
             elif line.startswith('vt '):
                 index_next.vt_index += 1
-                write_geo_line(line, geo_name, geo_lines)
+                write_geo_line(line, file_name, geo_lines)
 
-        geo_lines.pop('default')
         for name in geo_lines:
             write_geo_force(name, geo_lines)
 
-    return geo2material
+    return file2material
 
 
 def get_transform(array: list) -> np.ndarray:
@@ -315,7 +305,7 @@ def convert_camera(converter_info: ConverterInfo, camera: dict) -> (str, dict):
         }
     }
     converter_info.camera_names.add(name)
-    logger.info(f'Camera {name} converted')
+    log_info(f'Camera {name} converted')
 
     return name, camera_dict
 
@@ -403,10 +393,10 @@ def convert_light(converter_info: ConverterInfo, light: dict, name: str) -> (str
             raise ValueError(f'Only one environment is supported, but {converter_info.environment_name} and {name} are found')
         converter_info.environment_name = name
     else:
-        logger.info(f'Light type {light["type"]} not supported')
+        log_info(f'Light type {light["type"]} not supported')
         return None, None
 
-    logger.info(f'Light {name} converted')
+    log_info(f'Light {name} converted')
 
     return name, light_dict
 
@@ -486,13 +476,13 @@ def convert_material(converter_info: ConverterInfo, material: dict, name: str, g
         mateial_map_geo[name] = material_dict
 
     converter_info.material_names.add(name)
-    logger.info(f'Material {name} converted')
+    log_info(f'Material {name} converted')
 
     return name, material_dict
 
 
 def convert_geometry(converter_info: ConverterInfo, geo_name: str) -> (ConverterInfo, str, dict):
-    logger.info(f'Start to convert geometry {geo_name}')
+    log_info(f'Start to convert geometry {geo_name}')
 
     check_dir(converter_info.input_json_dir / geo_name, converter_info.output_json_dir / geo_name)
     check_dir(converter_info.input_geo_dir / geo_name, converter_info.output_geo_dir / geo_name)
@@ -574,7 +564,7 @@ def convert_geometry(converter_info: ConverterInfo, geo_name: str) -> (Converter
         geo_dict[copy_name] = geo_obj_dict.copy()
         converter_info.shape_names.add(copy_name)
 
-    logger.info(f'Geometry {geo_name} converted')
+    log_info(f'Geometry {geo_name} converted')
 
     return converter_info, geo_name, geo_dict
 
@@ -637,7 +627,8 @@ def create_main_scene_file(converter_info: ConverterInfo):
 
 
 def disney2luisa(input_project_dir: Path, output_project_dir: Path):
-    shutil.rmtree(output_project_dir)
+    if output_project_dir.exists():
+        shutil.rmtree(output_project_dir)
     check_dir(input_project_dir, output_project_dir)
 
     input_json_dir = input_project_dir / 'json'
@@ -665,7 +656,7 @@ def disney2luisa(input_project_dir: Path, output_project_dir: Path):
         camera_dict = {name: camera_dict}
         with open(path_out / camera_file.name, 'w') as f:
             json.dump(camera_dict, f, indent=2)
-    logger.info('All cameras converted')
+    log_info('All cameras converted')
 
     # lights
     path_in = converter_info.input_json_dir / 'lights'
@@ -691,7 +682,7 @@ f   -4   -3   -2   -1
             lights_dict[name] = light_dict
         with open(path_out / light_file.name, 'w') as f:
             json.dump(lights_dict, f, indent=2)
-    logger.info('All lights converted')
+    log_info('All lights converted')
 
     # geometries
     for geo_name in geo_names:
@@ -705,9 +696,9 @@ f   -4   -3   -2   -1
     threads = []
     thread2geo_name = {}
     for geo_name in geo_names:
-        # # DEBUG
-        # if geo_name not in test_geo_names:
-        #     continue
+        # DEBUG
+        if geo_name not in test_geo_names:
+            continue
         # geometries
         thread_t = pool.apply_async(convert_geometry, args=(converter_info.copy(), geo_name))
         threads.append(thread_t)
@@ -733,7 +724,7 @@ f   -4   -3   -2   -1
             with open(path_out / f'{geo_name}.json', 'w') as f:
                 json.dump(geo_dict, f, indent=2)
             finished_threads += 1
-            logger.info(f'Thread {finished_threads}/{thread_count} finished. Geometry {geo_name} dumped')
+            log_info(f'Thread {finished_threads}/{thread_count} finished. Geometry {geo_name} dumped')
         threads = threads_next
         time_now = time.perf_counter()
         time_delta += time_now - time_last
@@ -741,26 +732,25 @@ f   -4   -3   -2   -1
         if time_delta >= print_every_seconds:
             time_delta = 0.0
             thread_names = [thread2geo_name[thread_t] for thread_t in threads]
-            logger.info(f'Thread {finished_threads}/{thread_count} finished. Unfinished threads: ', thread_names)
+            log_info(f'Thread {finished_threads}/{thread_count} finished. Unfinished threads: ', thread_names)
     pool.join()
-    logger.info('All geometries converted')
+    log_info('All geometries converted')
 
     # materials for DEBUG
     json.dump(converter_info.re_shape2material, open(converter_info.output_project_dir / 're_shape2material.json', 'w'), indent=2)
     json.dump(converter_info.material_map, open(converter_info.output_project_dir / 'material_map.json', 'w'), indent=2)
-    logger.info('Materials for DEBUG dumped')
+    log_info('Materials for DEBUG dumped')
 
     # textures
     copy_texture(input_project_dir / 'textures', converter_info.output_project_dir / 'textures')
-    logger.info('Textures copied')
+    log_info('Textures copied')
 
     create_main_scene_file(converter_info)
-    logger.info('Main scene file created')
+    log_info('Main scene file created')
 
 
 if __name__ == '__main__':
-    logger_init()
     if len(argv) == 3:
         disney2luisa(Path(argv[1]).absolute(), Path(argv[2]).absolute())
     else:
-        logger.info('Usage: disney2luisa.py <input.json> <output.luisa>')
+        log_info('Usage: disney2luisa.py <input.json> <output.luisa>')
